+++ tests/Feature/Console/ConsoleSourceIndexerTest.php
@@ -75,7 +75,7 @@ final class ConsoleSourceIndexerTest extends TestCase
             $converter,
             new TolerantPhpParser(new Parser()),
             new PhpCommentParser(),
-            new ConsoleDefinitionExtractor($delimiters),
+            new ConsoleDefinitionExtractor(),
             new ConsoleInvokableParameterExtractor($delimiters),
             new ConsoleInputReceiverResolver(new PhpCapturedReceiverResolver($delimiters)),
         ));